Refactoring: Impl로 분리
